backend/requirements.txt
Flask==2.3.5
flask-cors==4.0.0
openpyxl==3.1.3
pillow==10.0.0
gunicorn==21.2.0
